11-13

+++ 11-13-lecture/php-12-13/public/index.php
@@ -1,15 +1,21 @@
 <?php
-require_once "../app/core/Database.php";
+require_once "../app/models/Model.php";
 require_once "../app/models/User.php";
 require_once "../app/controllers/UserController.php";
 
 //set our env variables
 $env = parse_ini_file('../.env');
-require '../app/core/config.php';
+define('DBNAME', $env['DBNAME']);
+define('DBHOST', $env['DBHOST']);
+define('DBUSER', $env['DBUSER']);
+define('DBPASS', $env['DBPASS']);
 
 use app\controllers\UserController;
 
+//get uri without query strings
 $uri = strtok($_SERVER["REQUEST_URI"], '?');
+
+//get uri pieces
 $uriArray = explode("/", $uri);
 //0 = ""
 //1 = users
@@ -17,54 +23,63 @@ $uriArray = explode("/", $uri);
 
 
 //get all or a single user
-if ($uriArray[1] === 'users' && $_SERVER['REQUEST_METHOD'] === 'GET') {
-    $id = isset($uriArray[2]) ? intval($uriArray[2]) : null;
+if ($uriArray[1] === 'api' && $uriArray[2] === 'users' && $_SERVER['REQUEST_METHOD'] === 'GET') {
+    //only id
+    $id = isset($uriArray[3]) ? intval($uriArray[3]) : null;
     $userController = new UserController();
-    $userController->getUsers($id);
+
+    if ($id) {
+        $userController->getUserByID($id);
+    } else {
+        $userController->getAllUsers();
+    }
 }
 
 //save user
-if ($uri === '/users' && $_SERVER['REQUEST_METHOD'] === 'POST') {
+if ($uriArray[1] === 'api' && $uriArray[2] === 'users' && $_SERVER['REQUEST_METHOD'] === 'POST') {
     $userController = new UserController();
     $userController->saveUser();
 }
 
 //update user
-if ($uriArray[1] === 'users' && $_SERVER['REQUEST_METHOD'] === 'PUT') {
+if ($uriArray[1] === 'api' && $uriArray[2] === 'users' && $_SERVER['REQUEST_METHOD'] === 'PUT') {
     $userController = new UserController();
-    $id = isset($uriArray[2]) ? intval($uriArray[2]) : null;
+    $id = isset($uriArray[3]) ? intval($uriArray[3]) : null;
     $userController->updateUser($id);
 }
 
 //delete user
-if ($uriArray[1] === 'users' && $_SERVER['REQUEST_METHOD'] === 'DELETE') {
+if ($uriArray[1] === 'api' && $uriArray[2] === 'users' && $_SERVER['REQUEST_METHOD'] === 'DELETE') {
     $userController = new UserController();
-    $id = isset($uriArray[2]) ? intval($uriArray[2]) : null;
+    $id = isset($uriArray[3]) ? intval($uriArray[3]) : null;
     $userController->deleteUser($id);
 }
 
 //views
-if ($uri === '/users-view' && $_SERVER['REQUEST_METHOD'] === 'GET') {
-    $userController = new UserController();
-    $userController->usersView();
-}
 
-if ($uri === '/users-add-view' && $_SERVER['REQUEST_METHOD'] === 'GET') {
+
+if ($uri === '/users-add' && $_SERVER['REQUEST_METHOD'] === 'GET') {
     $userController = new UserController();
     $userController->usersAddView();
 }
 
-if ($uriArray[1] === 'users-update-view' && $_SERVER['REQUEST_METHOD'] === 'GET') {
+if ($uriArray[1] === 'users-update' && $_SERVER['REQUEST_METHOD'] === 'GET') {
     $userController = new UserController();
     $userController->usersUpdateView();
 }
 
-if ($uriArray[1] === 'users-delete-view' && $_SERVER['REQUEST_METHOD'] === 'GET') {
+if ($uriArray[1] === 'users-delete' && $_SERVER['REQUEST_METHOD'] === 'GET') {
     $userController = new UserController();
     $userController->usersDeleteView();
 }
 
+if ($uriArray[1] === '' && $_SERVER['REQUEST_METHOD'] === 'GET') {
+    $userController = new UserController();
+    $userController->usersView();
+}
+
 include '../public/assets/views/notFound.html';
+exit();
 
 ?>
